schema modulation/merge
